pyproject.toml
[tool.poetry]
name = "email-ai-app"
version = "0.1.0"
description = "AI-powered email management application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
celery = "^5.4.0"
redis = "^5.0.1"
flower = "^2.0.1"
pydantic = {extras = ["email"], version = "^2.5.2"}
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
openai = "^1.12.0"
prometheus-client = "^0.12.0"
starlette-prometheus = "^0.9.0"
slowapi = "^0.1.8"
prometheus-fastapi-instrumentator = "^6.1.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.40.0"}
aiosmtplib = "^2.0.2"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 